+++ tools/check_site_structure.php
@@ -21,6 +21,7 @@ $requiredSite = [
     'session.php',
     'person_repository.php',
     'header.php',
+    'icon.svg',
     'version',
 ];
 
@@ -39,6 +40,9 @@ $requiredRootTools = [
     'check_v090_schema.php',
     'check_person_management.php',
     'check_person_created_print.php',
+    'check_favicon.php',
+    'check_person_delete_role_ui.php',
+    'check_person_password_handout.php',
     'check_qr_code_handout.php',
     'check_reasons_soft_delete.php',
     'check_absence_manual_delete.php',